+++ packages/mercadopago/tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts", "src/vercel-kv.ts", "src/otel.ts"],
+  entry: ["src/index.ts", "src/vercel-kv.ts", "src/otel.ts", "src/testing.ts"],
   external: ["@vercel/kv", "@opentelemetry/api"],
   format: ["esm", "cjs"],
   dts: true,